backend/requirements.txt
# MigrateMate Backend Dependencies

# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3

# Database (PostgreSQL + Async)
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Vector Database (Qdrant)
qdrant-client==1.9.1

# Graph Database (Neo4j)
neo4j==5.20.0

# Cache (Redis)
redis==5.0.4

# AI/ML
torch==2.3.0
sentence-transformers==3.0.0
langchain==0.2.5
langgraph==0.1.5
langchain-core==0.2.10

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
deepeval==0.21.0

# Development
python-dotenv==1.0.1
black==24.4.2
ruff==0.4.9

# Utilities
python-multipart==0.0.9
aiofiles==23.2.1

# Google Gemini
google-generativeai==0.5.4